+++ dev/docs/ReleaseNotes.md
@@ -1,5 +1,9 @@
 # Currency Speculator - Release Notes
 
+## 2020-05-15 v0.2
+#### [Story: Test Margin Fx Speculator]
+- **Add Feature**
+
 ## 2020-05-08 v0.1
 #### [Story: Immediate Margin Fx Speculator]
 - **Add Feature**
@@ -7,4 +11,4 @@
         - Calculate Immediate Margin
         - Using maximum 1 iterations
     - Class DelayedMarginFxSpeculator
-       - Calculate Margin over a period of 3 consecutive weeks    
+       - Calculate Margin over a period of 3 consecutive weeks